+++ src/WhereGroup/MetadorBundle/Component/XmlParser.php
@@ -158,8 +158,13 @@ class XmlParser
                     if (is_object($val)) {
                         $tmp = $this->parseRecursive($val, $key, $context, $path, $commands);
                     } elseif (is_array($val) && count($val) >= 2) {
-                        $tmp = $this->getValue($path . $val[0], $context);
-                        $tmp = $this->functions->get($val[1], $tmp, array_slice($val, 2));
+
+                        if (is_null($val[0])) {
+                            $tmp = $this->functions->get($val[1], null, array_slice($val, 2));
+                        } else {
+                            $tmp = $this->getValue($path . $val[0], $context);
+                            $tmp = $this->functions->get($val[1], $tmp, array_slice($val, 2));
+                        }
                     } else {
                         $tmp = $this->getValue($path.$val, $context);
                     }
@@ -239,6 +244,9 @@ class XmlParser
                         break;
                     case XML_TEXT_NODE:
                         $result[] = $node->wholeText;
+                        break;
+                    case XML_CDATA_SECTION_NODE:
+                        $result[] = $node->textContent;
                         break;
                     default:
                 }