+++ app/Resources/doc/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.6.0
+* Mapbender3 search integration.
+* License changed to MIT.
+* WMS GetCapabilities upload.
+* Added UUID component.
+* Search element for mapbender.
+* Added Paging.
+
 ## 2.0.5.0
 * XML-Import command.
 * Added build script.